requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-documentai==2.31.0
google-cloud-vision==3.7.3
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.65.1
idna==3.7
isodate==0.6.1
msrest==0.7.1
numpy==2.0.1
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pytesseract==0.3.10
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
